pyproject.toml
[project]
name = "minSDXL"
version = "0.1.0"
description = "Pure PyTorch SDXL fine-tuning with LoRA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "safetensors>=0.4.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "huggingface-hub>=0.19.0",
    "diffusers>=0.35.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sdxl*", "training*", "sampling*", "utils*"]
exclude = ["minSDXL*", "diffusers*", "test_data*", "*.tests*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
